requirements.txt
mss==9.0.1
numpy==1.25.2
opencv-python==4.8.0.74
pynput==1.7.6
pytesseract==0.3.10
python-vlc==3.0.18122
